Try and clean up some bad unicode stuff.
Try to decode all PPDs as UTF-8. This is still seems wrong. It should use the
LanguageEncoding in the PPD but a lot of UTF-8 PPDs seem to say they are
ISOLatin1. Not sure what the best way to handle this is.

Lets tell google we are sending them UTF-8 and ask for UTF-8 back.