Establish sasl authentication in alternate auth file if auth not established.
The alternate auth file is currently defined to support a format
with the username on the first line, and the password on the second.
This allows it to support the more secure sasl format used by the
default auth file.

If the alternate auth file is specified and authentication fails, this
will lead to the applications prompting for credentials, which are
ultimately stored in the sasl form.