Add xmpp error check to _check_for_notification()
It looks like xmpp msg()'s can result in more than one clause
returned by the host, and that only one is processed per msg()
call. If so, it is possible that a failure may not be processed
until the notification call, which does not check.

Add a failure check to the notification method.