Skip over printers with no PPD when trying to auth for the first time.
If we try to auth with a printer with no PPD, we end up getting:

> cups.IPPError: (1030, u'Not Found')

When we try to call the .ppd endpoint on the CUPS server (getPPD(printer_name)), which causes us to die with a non-useful error message.

With this change we will skip over any printers with no PPD file and auth with the first one that has one.

Fixes #130